Make Unicode size Py_ssize_t.
git-svn-id: http://svn.python.org/projects/python/branches/ssize_t@41781 6015fed2-1504-0410-9fe1-9d1591cc4771

+++ Include/unicodeobject.h
@@ -372,7 +372,7 @@ extern "C" {
 
 typedef struct {
     PyObject_HEAD
-    int length;			/* Length of raw Unicode data in buffer */
+    Py_ssize_t length;		/* Length of raw Unicode data in buffer */
     Py_UNICODE *str;		/* Raw Unicode buffer */
     long hash;			/* Hash value; -1 if not set */
     PyObject *defenc;		/* (Default) Encoded version as Python
